requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Pillow==10.4.0
WTForms==3.1.2
email-validator==2.2.0
psycopg[binary]==3.3.2
python-dotenv==1.0.1
